// src/shared/waygoal-types.ts
export interface WaygoalPoint { x: number; y: number }
export interface WaygoalView { x: number; y: number; scale: number }